Skip EventHUp notify in FIN_WAIT2 on a socket close.
This Notify was added as part of cl/279106406 but, notifying EventHUp
in FIN_WAIT2 is incorrect, as we want to only notify later on TIME_WAIT
or a reset. But, we do need to notify any blocked waiters of an
activity on the endpoint with EventIn|EventOut.
third_party/gvisor/pkg/tcpip/adapters/gonet:gonet_test:TestCloseReader
covers this case.

PiperOrigin-RevId: 339926447